Finding Prime Factors of a Number in C++
Prime factors are basically the factors of the number input which are prime themselves. For instance the prime factors of “24”, without duplication, are [ 12, 2, 6, 3] but the Prime Factors, without duplication, are [ 2, 3]. The best way to find out the prime factors is to find the factors of the given number input and check whether they’re prime numbers or not.